Cleaned up install scripts

+++ install.sh
@@ -21,8 +21,7 @@ download () {
 
     # Pull ecTrans.
     info "==> PULLING ECTRANS"
-    cd ${SOURCEDIR}
-    # git clone --branch alaro_ectrans https://github.com/ddegrauwe/ectrans.git
+    cd "${SOURCEDIR}" || exit 1
     git clone https://github.com/OkkeVanEck/ectrans.git
     success "==> SUCCESFULLY CLONED ECTRANS"
 
@@ -39,16 +38,17 @@ download () {
 _build_install_ecbuild () {
     # Build and Install ecBuild.
     info "==> INSTALLING ECBUILD.."
-    cd ${SOURCEDIR}/ecbuild
+    cd "${SOURCEDIR}/ecbuild" || exit 1
 
     # Remove obsolete switch '-Gfast'.
     sed -i -e "s/-Gfast//" cmake/compiler_flags/Cray_Fortran.cmake
     
     # Create build directory and build ecBuild.
-    mkdir -p ${BUILDDIR}/ecbuild
-    cd ${BUILDDIR}/ecbuild
+    mkdir -p "${BUILDDIR}/ecbuild"
+    cd "${BUILDDIR}/ecbuild" || exit 1
     info "==>\t ECBUILD.."
-    ${SOURCEDIR}/ecbuild/bin/ecbuild --prefix=${INSTALLDIR}/ecbuild ${SOURCEDIR}/ecbuild
+    "${SOURCEDIR}/ecbuild/bin/ecbuild" --prefix="${INSTALLDIR}/ecbuild" \
+        "${SOURCEDIR}/ecbuild"
     info "==>\t MAKE.."
     make
 
@@ -62,16 +62,18 @@ _build_install_ecbuild () {
 _build_install_fiat () {
     # Build and Install FIAT.
     info "==> INSTALLING FIAT.."
-    cd ${SOURCEDIR}/fiat
-    # small fix to include OpenMP in linking of C programs
+    cd "${SOURCEDIR}/fiat" || exit 1
+    # Small fix to include OpenMP in linking of C programs
     sed -i -e "s/target_link_libraries( fiat-printbinding OpenMP::OpenMP_C )/target_link_libraries( fiat-printbinding \$\{OpenMP_C_FLAGS\} OpenMP::OpenMP_C )/" src/programs/CMakeLists.txt
     
     # Create build directory and build FIAT.
-    rm -rf ${BUILDDIR}/${FIAT_DIR}
-    mkdir -p ${BUILDDIR}/${FIAT_DIR}
-    cd ${BUILDDIR}/${FIAT_DIR}
+    rm -rf "${BUILDDIR:?}/${FIAT_DIR:?}"
+    mkdir -p "${BUILDDIR}/${FIAT_DIR}"
+    cd "${BUILDDIR}/${FIAT_DIR}" || exit 1
     info "==>\t ECBUILD.."
-    ecbuild -DCMAKE_BUILD_TYPE=RELEASE -DCMAKE_INSTALL_PREFIX=${INSTALLDIR}/${FIAT_DIR} -DENABLE_MPI=ON -DBUILD_SHARED_LIBS=OFF -DENABLE_TESTS=OFF ${SOURCEDIR}/fiat
+    ecbuild -DCMAKE_BUILD_TYPE=RELEASE \
+        -DCMAKE_INSTALL_PREFIX="${INSTALLDIR}/${FIAT_DIR}" -DENABLE_MPI=ON \
+        -DBUILD_SHARED_LIBS=OFF -DENABLE_TESTS=OFF "${SOURCEDIR}/fiat"
     info "==>\t MAKE.."
     make -j16
 
@@ -87,11 +89,17 @@ _build_install_ectrans () {
     info "==> Installing ecTrans.."
 
     # Create build directory and build ecTrans.
-    rm -rf ${BUILDDIR}/${ECTRANS_DIR} ${INSTALLDIR}/${ECTRANS_DIR}
-    mkdir -p ${BUILDDIR}/${ECTRANS_DIR}
-    cd ${BUILDDIR}/${ECTRANS_DIR}
+    rm -rf "${BUILDDIR:?}/${ECTRANS_DIR:?}" "${INSTALLDIR:?}/${ECTRANS_DIR:?}"
+    mkdir -p "${BUILDDIR}/${ECTRANS_DIR}"
+    cd "${BUILDDIR}/${ECTRANS_DIR}" || exit 1
     info "==>\t ECBUILD.."
-    ecbuild --prefix=${INSTALLDIR}/${ECTRANS_DIR} -DCMAKE_BUILD_TYPE=RelWithDebInfo -Dfiat_ROOT=${INSTALLDIR}/${FIAT_DIR} -DBUILD_SHARED_LIBS=OFF -DENABLE_FFTW=OFF -DENABLE_GPU=ON -DENABLE_OMPGPU=OFF -DENABLE_ACCGPU=ON -DENABLE_TESTS=OFF -DENABLE_GPU_AWARE_MPI=ON -DENABLE_CPU=ON -DENABLE_ETRANS=ON  -DENABLE_DOUBLE_PRECISION=ON -DENABLE_SINGLE_PRECISION=OFF ${SOURCEDIR}/ectrans
+    ecbuild --prefix="${INSTALLDIR}/${ECTRANS_DIR}" \
+        -DCMAKE_BUILD_TYPE=RelWithDebInfo \
+        -Dfiat_ROOT="${INSTALLDIR}/${FIAT_DIR}" -DBUILD_SHARED_LIBS=OFF \
+        -DENABLE_FFTW=OFF -DENABLE_GPU=ON -DENABLE_OMPGPU=OFF \
+        -DENABLE_ACCGPU=ON -DENABLE_TESTS=OFF -DENABLE_GPU_AWARE_MPI=ON \
+        -DENABLE_CPU=ON -DENABLE_ETRANS=ON  -DENABLE_DOUBLE_PRECISION=ON \
+        -DENABLE_SINGLE_PRECISION=OFF "${SOURCEDIR}/ectrans"
     info "==>\t MAKE (supposed to fail).."
     make LIBRARY_PATH=/opt/cray/pe/cce/16.0.1/cce-clang/x86_64/lib -j32
     # Make needs to be executed twice as the first time it somehow fails.
@@ -135,7 +143,7 @@ main () {
     else
         # Else parse every passed argument.
         for var in "$@"; do
-            arrVar=(${var//:/ })
+            arrVar=("${var//:/ }")
             instruction=${arrVar[0]}
             program=${arrVar[1]}
